Modularizado el schema y funciona 10/10
