+++ graphql/resolvers/patientResolver.js
@@ -1,4 +1,4 @@
-const Patient = require('../models/Patient');
+const Patient = require('../../models/Patient');
 
 module.exports={
     Query: {
