+++ graphql/types/Date.js
@@ -1,3 +1,3 @@
-export default`
+module.exports=`
 scalar Date
 `;
