+++ graphql/types/Institute.js
@@ -1,5 +1,9 @@
-export default`
+module.exports=`
 type Institute{
     _id:ID!,
     institute_name:String
-}`;
+}
+type Query{
+    institutes: [Institute]
+}
+`;
