+++ graphql/types/Institute_Last_Revision.js
@@ -1,4 +1,4 @@
-export default `
+module.exports= `
 type Institute_Last_Revision{
     _id: ID!,
     _last_revision: String,
