+++ graphql/types/User.js
@@ -1,4 +1,4 @@
-export default
+module.exports=
 `type User{
     _id: ID!,
     username: String,
