+++ graphql/types/index.js
@@ -3,4 +3,4 @@ const {fileLoader, mergeTypes}=require('merge-graphql-schemas');
 
 const typesArray=fileLoader(path.join(__dirname,'./'));
 
-export default mergeTypes(typesArray, {all:true});
+module.exports= mergeTypes(typesArray, {all:true});
